Inventory Management System.
This is an Inventory Management System. This is basically a marketing project which have all the products and generates a bill after successful shipping. As it has three files sales.json, record.json and the source code file. The files get automatically updated after every purchase. In this way, this is an interesting program and project.