Password Strength Checker.
This is the program to check the strength of the passwords. As we need to create strong passwords to create account on various websites and apps. So it is necessary that the password chosen should be strong enough so that it can not be easily predicted. This program solves this problem as it tells the strength of the chosen password and one can modify it if necessary based on the generated results.